+++ src/constants/constants.js
@@ -1,4 +1,5 @@
-export const MAX_NUMBER_LIMIT = 1000000;
+// Limit
+export const LIMIT_MAX_NUMBER = 1000000;
 
 // Constants to define rules of words forming for cardinal numbers
 export const CARDINAL_NUMBER_BASIC_MAX = 9;
@@ -19,3 +20,11 @@ export const THOUSAND_STRING = 'thousand';
 export const MILLION_STRING = 'million';
 export const BILLION_STRING = 'billion';
 export const TRILLION_STRING = 'trillion';
+
+// Regex
+// mask to put commas into number: 111222333 -> 111,222,333
+export const REGEX_THREE_DIGITS_CHUNK = /\B(?=(\d{3})+(?!\d))/g;
+// mask of numbers with commas: 6,807; 143,562; 12,008 etc.
+export const REGEX_COMMAS_ORDER = /^(?:\d{1,3}(?:,\d{3})*|\d+)$/;
+// mask of only digits and comma in input: 6807; 143,562; 12008 etc.
+export const REGEX_SUPPORTED_SYMBOLS = /^[,0-9]+$/;