+++ CHANGELOG.md
@@ -2,9 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.3.0 / 2022-07-03
+
+### Added
+
+- Unit tests covering main NumberToWord method;
+- Unit tests covering Input validation method;
+- Updated README with test running instruction.
+
+### Fixed
+
+- Added separated logger module;
+- Extracted logic of input validation into separate method;
+- Fixed structure of the project.
+
 ## 1.2.0 / 2022-07-02
 
-* Added option to switch off the limit
+### Added
+
+- Option to switch off the default limit;
+- Updated README with option usage instruction.
+
+### Fixed
+
+- Updated number parsing logic.
 
 ## 1.1.0 / 2022-07-01